refactor: label payload logs with [payload] marker and extract url_sanitization module #436
Address review feedback on #459:
- log_payload prefixes every record with a [payload] marker so payload
  lines can be filtered with a single query
- sanitize_url_for_log moves from url_classification to a dedicated
  common/url_sanitization module (log hygiene, not classification)
