+++ src/quickapp/dial_core_services/dial_downloader.py
@@ -4,7 +4,7 @@ from aidial_client import AsyncDial
 from aidial_client.types.metadata import FileMetadata
 from injector import inject
 
-from quickapp.common.url_classification import sanitize_url_for_log
+from quickapp.common.url_sanitization import sanitize_url_for_log
 from quickapp.shared.config_resolvers.file_loading_size_limit_resolver import (
     FileLoadingSizeLimitResolver,
 )
