+++ src/quickapp/dial_files_tooling/_path_argument_transformer.py
@@ -4,7 +4,7 @@ from typing import Any
 from injector import inject
 
 from quickapp.common.abstract.base_tool_argument_transformer import ToolArgumentTransformer
-from quickapp.common.url_classification import sanitize_url_for_log
+from quickapp.common.url_sanitization import sanitize_url_for_log
 from quickapp.dial_files_tooling._home_path_resolver import _HomePathResolver
 
 logger = logging.getLogger(__name__)
